fix(lsp): remove action.go file and all associated methods
refactor(lsp): consolidate CancelRequest into requests.go

fix(imports): clean up unused imports from cancel.go

refactor(lsp): move TextDocumentDidChangeNotification to notifications.go

refactor(lsp): relocate cancel and shutdown methods to requests.go

feat(lsp): add ShutdownRequest and ShutdownResponse methods to requests.go

fix(lsp): delete change.go file and merge its methods into notifications.go

cleanup(lsp): delete redundant textDocumentCompletion.go file

cleanup(lsp): delete textDocumentDidSave.go file and related content
