--- tools/seltabls/pkg/lsp/action.go
@@ -1,1 +1,0 @@
-package lsp
