+++ tools/seltabls/pkg/lsp/responses.go
@@ -114,3 +114,33 @@ func NewInitializeResponse(
 		}
 	}
 }
+
+// ShutdownResponse is the response to a ShutdownRequest.
+//
+// Microsoft LSP Docs:
+// https://microsoft.github.io/language-server-protocol/specifications/specification-current/#shutdown
+type ShutdownResponse struct {
+	Response
+	Result *bool  `json:"result"`
+	Error  *error `json:"error,omitempty"`
+}
+
+// Method returns the method for the shutdown response
+func (r ShutdownResponse) Method() methods.Method {
+	return methods.MethodShutdown
+}
+
+// NewShutdownResponse creates a new shutdown response
+func NewShutdownResponse(
+	request ShutdownRequest,
+	err error,
+) (ShutdownResponse, error) {
+	return ShutdownResponse{
+		Response: Response{
+			RPC: RPCVersion,
+			ID:  request.ID,
+		},
+		Result: nil,
+		Error:  &err,
+	}, nil
+}
