--- tools/seltabls/pkg/lsp/textDocumentCompletion.go
@@ -1,1 +1,0 @@
-package lsp
