--- tools/seltabls/pkg/lsp/textDocumentDidSave.go
@@ -1,1 +1,0 @@
-package lsp
